requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mysqlclient==2.2.0
django-cors-headers==4.3.0
django-filter==23.3
django-crispy-forms==2.3
crispy-bootstrap5==2025.4
django-allauth==0.58.2
django-bootstrap5==23.3
drf-yasg==1.21.7
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
xlsxwriter==3.1.2
reportlab==4.0.9
